+++ test/widget_test.dart
@@ -5,32 +5,26 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:study_riverpod/main.dart';
 
 void main() {
-  testWidgets('Widget動作テスト', (WidgetTester tester) async {
-    // Widgetの構築
-    await tester.pumpWidget(
-      const ProviderScope(
-        child: MyApp(),
-      ),
-    );
+  testWidgets('Counter increments smoke test', (WidgetTester tester) async {
+    // Build our app and trigger a frame.
+    await tester.pumpWidget(const MyApp());
 
-    // 初期表示
+    // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
+    expect(find.text('1'), findsNothing);
 
-    // 「+」を2回クリックした後の確認
-    await tester.tap(find.text('+'));
-    await tester.tap(find.text('+'));
+    // Tap the '+' icon and trigger a frame.
+    await tester.tap(find.byIcon(Icons.add));
     await tester.pump();
-    expect(find.text('2'), findsOneWidget);
 
-    // 「-」を1回クリックした後の確認
-    await tester.tap(find.text('-'));
-    await tester.pump();
+    // Verify that our counter has incremented.
+    expect(find.text('0'), findsNothing);
     expect(find.text('1'), findsOneWidget);
   });
 }
